refactor: auto-split test_series.py
