+++ tests/shared/python/signal_toolkit/series/test_coefficients.py
@@ -1,0 +1,127 @@
+"""Tests for the Taylor and Maclaurin series module.
+
+This module contains comprehensive tests for series expansion functionality:
+- Taylor series computation at any point
+- Maclaurin series (Taylor at x=0)
+- Common function series (sin, cos, exp, ln, etc.)
+- Convergence analysis
+- Error bounds
+
+Following TDD and Design by Contract principles.
+"""
+
+from __future__ import annotations
+
+import math
+
+import numpy as np
+import pytest
+
+# =============================================================================
+# SeriesExpansion Class Contract Tests
+# =============================================================================
+
+
+# =============================================================================
+# Taylor Series Method Contract Tests
+# =============================================================================
+
+
+# =============================================================================
+# Maclaurin Series Contract Tests
+# =============================================================================
+
+
+# =============================================================================
+# Get Coefficients Contract Tests
+# =============================================================================
+
+
+# =============================================================================
+# Common Series Functions Contract Tests
+# =============================================================================
+
+
+# =============================================================================
+# Taylor Series Functional Tests
+# =============================================================================
+
+
+# =============================================================================
+# Common Series Functions Tests
+# =============================================================================
+
+
+# =============================================================================
+# Coefficients Tests
+# =============================================================================
+
+
+class TestCoefficients:
+    """Tests for Taylor series coefficient extraction."""
+
+    def test_exp_coefficients(self) -> None:
+        """Test exponential series coefficients are approximately 1/n!."""
+        from signal_toolkit.series import SeriesExpansion
+
+        expansion = SeriesExpansion()
+        coeffs = expansion.get_coefficients(np.exp, center=0, n_terms=10)
+
+        expected = [1 / math.factorial(n) for n in range(10)]
+        # Polynomial fitting gives good approximations but not exact values
+        # The key test is that the resulting series produces accurate results
+        # Higher-order coefficients have more numerical error
+        assert np.allclose(coeffs, expected, rtol=5e-2)
+
+    def test_sin_coefficients(self) -> None:
+        """Test sine series coefficients are approximately correct."""
+        from signal_toolkit.series import SeriesExpansion
+
+        expansion = SeriesExpansion()
+        coeffs = expansion.get_coefficients(np.sin, center=0, n_terms=7)
+
+        # sin(x) = x - x^3/3! + x^5/5! - ...
+        # Coefficients: 0, 1, 0, -1/6, 0, 1/120, 0
+        expected = [0, 1, 0, -1 / 6, 0, 1 / 120, 0]
+        # Polynomial fitting gives good approximations; zero coefficients may have
+        # small numerical noise
+        assert np.allclose(coeffs, expected, rtol=1e-2, atol=1e-3)
+
+    def test_cos_coefficients(self) -> None:
+        """Test cosine series coefficients are approximately correct."""
+        from signal_toolkit.series import SeriesExpansion
+
+        expansion = SeriesExpansion()
+        coeffs = expansion.get_coefficients(np.cos, center=0, n_terms=7)
+
+        # cos(x) = 1 - x^2/2! + x^4/4! - ...
+        # Coefficients: 1, 0, -1/2, 0, 1/24, 0, -1/720
+        expected = [1, 0, -0.5, 0, 1 / 24, 0, -1 / 720]
+        # Polynomial fitting gives good approximations; zero coefficients may have
+        # small numerical noise
+        assert np.allclose(coeffs, expected, rtol=1e-2, atol=1e-3)
+
+
+# =============================================================================
+# Convergence Analysis Tests
+# =============================================================================
+
+
+# =============================================================================
+# Error Bounds Tests
+# =============================================================================
+
+
+# =============================================================================
+# SeriesResult Dataclass Tests
+# =============================================================================
+
+
+# =============================================================================
+# Integration with Signal Toolkit Tests
+# =============================================================================
+
+
+# =============================================================================
+# Utility Functions Tests
+# =============================================================================
