+++ tests/shared/python/signal_toolkit/series/test_common_series_contract.py
@@ -1,0 +1,134 @@
+"""Tests for the Taylor and Maclaurin series module.
+
+This module contains comprehensive tests for series expansion functionality:
+- Taylor series computation at any point
+- Maclaurin series (Taylor at x=0)
+- Common function series (sin, cos, exp, ln, etc.)
+- Convergence analysis
+- Error bounds
+
+Following TDD and Design by Contract principles.
+"""
+
+from __future__ import annotations
+
+import math
+
+import numpy as np
+import pytest
+
+# =============================================================================
+# SeriesExpansion Class Contract Tests
+# =============================================================================
+
+
+# =============================================================================
+# Taylor Series Method Contract Tests
+# =============================================================================
+
+
+# =============================================================================
+# Maclaurin Series Contract Tests
+# =============================================================================
+
+
+# =============================================================================
+# Get Coefficients Contract Tests
+# =============================================================================
+
+
+# =============================================================================
+# Common Series Functions Contract Tests
+# =============================================================================
+
+
+class TestCommonSeriesContract:
+    """Design by Contract tests for pre-defined common series."""
+
+    def test_exp_series_exists(self) -> None:
+        """Postcondition: exp_series function exists and is callable."""
+        from signal_toolkit.series import exp_series
+
+        assert callable(exp_series)
+
+    def test_sin_series_exists(self) -> None:
+        """Postcondition: sin_series function exists and is callable."""
+        from signal_toolkit.series import sin_series
+
+        assert callable(sin_series)
+
+    def test_cos_series_exists(self) -> None:
+        """Postcondition: cos_series function exists and is callable."""
+        from signal_toolkit.series import cos_series
+
+        assert callable(cos_series)
+
+    def test_ln_series_exists(self) -> None:
+        """Postcondition: ln_series function exists (for ln(1+x))."""
+        from signal_toolkit.series import ln_series
+
+        assert callable(ln_series)
+
+    def test_geometric_series_exists(self) -> None:
+        """Postcondition: geometric_series function exists (for 1/(1-x))."""
+        from signal_toolkit.series import geometric_series
+
+        assert callable(geometric_series)
+
+    def test_arctan_series_exists(self) -> None:
+        """Postcondition: arctan_series function exists."""
+        from signal_toolkit.series import arctan_series
+
+        assert callable(arctan_series)
+
+    def test_sinh_series_exists(self) -> None:
+        """Postcondition: sinh_series function exists."""
+        from signal_toolkit.series import sinh_series
+
+        assert callable(sinh_series)
+
+    def test_cosh_series_exists(self) -> None:
+        """Postcondition: cosh_series function exists."""
+        from signal_toolkit.series import cosh_series
+
+        assert callable(cosh_series)
+
+
+# =============================================================================
+# Taylor Series Functional Tests
+# =============================================================================
+
+
+# =============================================================================
+# Common Series Functions Tests
+# =============================================================================
+
+
+# =============================================================================
+# Coefficients Tests
+# =============================================================================
+
+
+# =============================================================================
+# Convergence Analysis Tests
+# =============================================================================
+
+
+# =============================================================================
+# Error Bounds Tests
+# =============================================================================
+
+
+# =============================================================================
+# SeriesResult Dataclass Tests
+# =============================================================================
+
+
+# =============================================================================
+# Integration with Signal Toolkit Tests
+# =============================================================================
+
+
+# =============================================================================
+# Utility Functions Tests
+# =============================================================================
