+++ tests/shared/python/signal_toolkit/series/test_convergence_analysis.py
@@ -1,0 +1,136 @@
+"""Tests for the Taylor and Maclaurin series module.
+
+This module contains comprehensive tests for series expansion functionality:
+- Taylor series computation at any point
+- Maclaurin series (Taylor at x=0)
+- Common function series (sin, cos, exp, ln, etc.)
+- Convergence analysis
+- Error bounds
+
+Following TDD and Design by Contract principles.
+"""
+
+from __future__ import annotations
+
+import math
+
+import numpy as np
+import pytest
+
+# =============================================================================
+# SeriesExpansion Class Contract Tests
+# =============================================================================
+
+
+# =============================================================================
+# Taylor Series Method Contract Tests
+# =============================================================================
+
+
+# =============================================================================
+# Maclaurin Series Contract Tests
+# =============================================================================
+
+
+# =============================================================================
+# Get Coefficients Contract Tests
+# =============================================================================
+
+
+# =============================================================================
+# Common Series Functions Contract Tests
+# =============================================================================
+
+
+# =============================================================================
+# Taylor Series Functional Tests
+# =============================================================================
+
+
+# =============================================================================
+# Common Series Functions Tests
+# =============================================================================
+
+
+# =============================================================================
+# Coefficients Tests
+# =============================================================================
+
+
+# =============================================================================
+# Convergence Analysis Tests
+# =============================================================================
+
+
+class TestConvergenceAnalysis:
+    """Tests for convergence analysis functionality."""
+
+    def test_convergence_analysis_returns_dict(self) -> None:
+        """Postcondition: analyze_convergence returns a dictionary."""
+        from signal_toolkit.series import SeriesExpansion
+
+        expansion = SeriesExpansion()
+        analysis = expansion.analyze_convergence(np.exp, center=0, x_test=1.0)
+
+        assert isinstance(analysis, dict)
+
+    def test_convergence_analysis_has_required_keys(self) -> None:
+        """Postcondition: Analysis has all required keys."""
+        from signal_toolkit.series import SeriesExpansion
+
+        expansion = SeriesExpansion()
+        analysis = expansion.analyze_convergence(np.exp, center=0, x_test=1.0)
+
+        assert "convergent" in analysis
+        assert "terms_for_convergence" in analysis
+        assert "final_error" in analysis
+        assert "errors_by_term" in analysis
+
+    def test_exp_converges_everywhere(self) -> None:
+        """Test that exp series converges for x values near center."""
+        from signal_toolkit.series import SeriesExpansion
+
+        expansion = SeriesExpansion()
+
+        # For numerical methods, convergence is reliable close to the center
+        # Far from center (|x| > 2), many terms are needed
+        for x in [-1, 0, 1]:
+            analysis = expansion.analyze_convergence(
+                np.exp, center=0, x_test=x, tolerance=1e-6
+            )
+            assert analysis["convergent"]
+
+    def test_ln_diverges_outside_radius(self) -> None:
+        """Test that ln(1+x) series diverges for x < -1 or x > 1."""
+        from signal_toolkit.series import SeriesExpansion
+
+        expansion = SeriesExpansion()
+
+        def f(x) -> float | np.ndarray:
+            return np.log(1 + x)
+
+        # Should diverge for x = 2 (outside |x| < 1)
+        analysis = expansion.analyze_convergence(
+            f, center=0, x_test=2.0, tolerance=1e-6
+        )
+        assert not analysis["convergent"]
+
+
+# =============================================================================
+# Error Bounds Tests
+# =============================================================================
+
+
+# =============================================================================
+# SeriesResult Dataclass Tests
+# =============================================================================
+
+
+# =============================================================================
+# Integration with Signal Toolkit Tests
+# =============================================================================
+
+
+# =============================================================================
+# Utility Functions Tests
+# =============================================================================
