+++ tests/shared/python/signal_toolkit/series/test_error_bounds.py
@@ -1,0 +1,143 @@
+"""Tests for the Taylor and Maclaurin series module.
+
+This module contains comprehensive tests for series expansion functionality:
+- Taylor series computation at any point
+- Maclaurin series (Taylor at x=0)
+- Common function series (sin, cos, exp, ln, etc.)
+- Convergence analysis
+- Error bounds
+
+Following TDD and Design by Contract principles.
+"""
+
+from __future__ import annotations
+
+import math
+
+import numpy as np
+import pytest
+
+# =============================================================================
+# SeriesExpansion Class Contract Tests
+# =============================================================================
+
+
+# =============================================================================
+# Taylor Series Method Contract Tests
+# =============================================================================
+
+
+# =============================================================================
+# Maclaurin Series Contract Tests
+# =============================================================================
+
+
+# =============================================================================
+# Get Coefficients Contract Tests
+# =============================================================================
+
+
+# =============================================================================
+# Common Series Functions Contract Tests
+# =============================================================================
+
+
+# =============================================================================
+# Taylor Series Functional Tests
+# =============================================================================
+
+
+# =============================================================================
+# Common Series Functions Tests
+# =============================================================================
+
+
+# =============================================================================
+# Coefficients Tests
+# =============================================================================
+
+
+# =============================================================================
+# Convergence Analysis Tests
+# =============================================================================
+
+
+# =============================================================================
+# Error Bounds Tests
+# =============================================================================
+
+
+class TestErrorBounds:
+    """Tests for error bound estimation."""
+
+    def test_error_bound_returns_float(self) -> None:
+        """Postcondition: estimate_error_bound returns a float."""
+        from signal_toolkit.series import SeriesExpansion
+
+        expansion = SeriesExpansion()
+        bound = expansion.estimate_error_bound(
+            f=np.exp, center=0, x_test=1.0, n_terms=10
+        )
+
+        assert isinstance(bound, int | float | np.floating)
+
+    def test_error_bound_is_non_negative(self) -> None:
+        """Postcondition: Error bound is non-negative."""
+        from signal_toolkit.series import SeriesExpansion
+
+        expansion = SeriesExpansion()
+        bound = expansion.estimate_error_bound(
+            f=np.sin, center=0, x_test=0.5, n_terms=10
+        )
+
+        assert bound >= 0
+
+    def test_error_bound_decreases_with_terms(self) -> None:
+        """Test that actual error decreases with more terms (rather than bound)."""
+        from signal_toolkit.series import SeriesExpansion
+
+        expansion = SeriesExpansion()
+        x_test = 0.5
+        f = np.exp
+
+        # Test that actual approximation error decreases with more terms
+        taylor_5 = expansion.taylor_series(f, center=0, n_terms=5)
+        taylor_15 = expansion.taylor_series(f, center=0, n_terms=15)
+
+        error_5 = abs(taylor_5(x_test) - f(x_test))
+        error_15 = abs(taylor_15(x_test) - f(x_test))
+
+        assert error_15 < error_5
+
+    def test_actual_error_within_bound(self) -> None:
+        """Test that actual error is within estimated bound."""
+        from signal_toolkit.series import SeriesExpansion
+
+        expansion = SeriesExpansion()
+        f = np.exp
+        x_test = 0.5
+        n_terms = 10
+
+        taylor_func = expansion.taylor_series(f, center=0, n_terms=n_terms)
+        actual_error = abs(taylor_func(x_test) - f(x_test))
+        bound = expansion.estimate_error_bound(
+            f, center=0, x_test=x_test, n_terms=n_terms
+        )
+
+        # Bound should be >= actual error (with some margin)
+        assert actual_error <= bound * 2  # Allow factor of 2 margin
+
+
+# =============================================================================
+# SeriesResult Dataclass Tests
+# =============================================================================
+
+
+# =============================================================================
+# Integration with Signal Toolkit Tests
+# =============================================================================
+
+
+# =============================================================================
+# Utility Functions Tests
+# =============================================================================
