+++ tests/shared/python/signal_toolkit/series/test_get_coefficients_contract.py
@@ -1,0 +1,112 @@
+"""Tests for the Taylor and Maclaurin series module.
+
+This module contains comprehensive tests for series expansion functionality:
+- Taylor series computation at any point
+- Maclaurin series (Taylor at x=0)
+- Common function series (sin, cos, exp, ln, etc.)
+- Convergence analysis
+- Error bounds
+
+Following TDD and Design by Contract principles.
+"""
+
+from __future__ import annotations
+
+import math
+
+import numpy as np
+import pytest
+
+# =============================================================================
+# SeriesExpansion Class Contract Tests
+# =============================================================================
+
+
+# =============================================================================
+# Taylor Series Method Contract Tests
+# =============================================================================
+
+
+# =============================================================================
+# Maclaurin Series Contract Tests
+# =============================================================================
+
+
+# =============================================================================
+# Get Coefficients Contract Tests
+# =============================================================================
+
+
+class TestGetCoefficientsContract:
+    """Design by Contract tests for get_coefficients method."""
+
+    def test_series_returns_array(self) -> None:
+        """Postcondition: Returns numpy array of coefficients."""
+        from signal_toolkit.series import SeriesExpansion
+
+        expansion = SeriesExpansion()
+
+        def f(x) -> float | np.ndarray:
+            return np.exp(x)
+
+        coeffs = expansion.get_coefficients(f, center=0, n_terms=5)
+
+        assert isinstance(coeffs, np.ndarray)
+
+    def test_correct_number_of_coefficients(self) -> None:
+        """Postcondition: Returns exactly n_terms coefficients."""
+        from signal_toolkit.series import SeriesExpansion
+
+        expansion = SeriesExpansion()
+
+        def f(x) -> float | np.ndarray:
+            return np.sin(x)
+
+        for n in [3, 5, 10]:
+            coeffs = expansion.get_coefficients(f, center=0, n_terms=n)
+            assert len(coeffs) == n
+
+
+# =============================================================================
+# Common Series Functions Contract Tests
+# =============================================================================
+
+
+# =============================================================================
+# Taylor Series Functional Tests
+# =============================================================================
+
+
+# =============================================================================
+# Common Series Functions Tests
+# =============================================================================
+
+
+# =============================================================================
+# Coefficients Tests
+# =============================================================================
+
+
+# =============================================================================
+# Convergence Analysis Tests
+# =============================================================================
+
+
+# =============================================================================
+# Error Bounds Tests
+# =============================================================================
+
+
+# =============================================================================
+# SeriesResult Dataclass Tests
+# =============================================================================
+
+
+# =============================================================================
+# Integration with Signal Toolkit Tests
+# =============================================================================
+
+
+# =============================================================================
+# Utility Functions Tests
+# =============================================================================
