+++ tests/shared/python/signal_toolkit/series/test_maclaurin_series_contract.py
@@ -1,0 +1,114 @@
+"""Tests for the Taylor and Maclaurin series module.
+
+This module contains comprehensive tests for series expansion functionality:
+- Taylor series computation at any point
+- Maclaurin series (Taylor at x=0)
+- Common function series (sin, cos, exp, ln, etc.)
+- Convergence analysis
+- Error bounds
+
+Following TDD and Design by Contract principles.
+"""
+
+from __future__ import annotations
+
+import math
+
+import numpy as np
+import pytest
+
+# =============================================================================
+# SeriesExpansion Class Contract Tests
+# =============================================================================
+
+
+# =============================================================================
+# Taylor Series Method Contract Tests
+# =============================================================================
+
+
+# =============================================================================
+# Maclaurin Series Contract Tests
+# =============================================================================
+
+
+class TestMaclaurinSeriesContract:
+    """Design by Contract tests for maclaurin_series method (Taylor at x=0)."""
+
+    def test_returns_callable(self) -> None:
+        """Postcondition: Returns a callable function."""
+        from signal_toolkit.series import SeriesExpansion
+
+        expansion = SeriesExpansion()
+
+        def f(x) -> float | np.ndarray:
+            return np.sin(x)
+
+        maclaurin_func = expansion.maclaurin_series(f, n_terms=5)
+
+        assert callable(maclaurin_func)
+
+    def test_equivalent_to_taylor_at_zero(self) -> None:
+        """Postcondition: Maclaurin series equals Taylor series at center=0."""
+        from signal_toolkit.series import SeriesExpansion
+
+        expansion = SeriesExpansion()
+
+        def f(x) -> float | np.ndarray:
+            return np.exp(x)
+
+        taylor_func = expansion.taylor_series(f, center=0, n_terms=10)
+        maclaurin_func = expansion.maclaurin_series(f, n_terms=10)
+
+        x_test = np.linspace(-1, 1, 20)
+        assert np.allclose(taylor_func(x_test), maclaurin_func(x_test))
+
+
+# =============================================================================
+# Get Coefficients Contract Tests
+# =============================================================================
+
+
+# =============================================================================
+# Common Series Functions Contract Tests
+# =============================================================================
+
+
+# =============================================================================
+# Taylor Series Functional Tests
+# =============================================================================
+
+
+# =============================================================================
+# Common Series Functions Tests
+# =============================================================================
+
+
+# =============================================================================
+# Coefficients Tests
+# =============================================================================
+
+
+# =============================================================================
+# Convergence Analysis Tests
+# =============================================================================
+
+
+# =============================================================================
+# Error Bounds Tests
+# =============================================================================
+
+
+# =============================================================================
+# SeriesResult Dataclass Tests
+# =============================================================================
+
+
+# =============================================================================
+# Integration with Signal Toolkit Tests
+# =============================================================================
+
+
+# =============================================================================
+# Utility Functions Tests
+# =============================================================================
