+++ tests/shared/python/signal_toolkit/series/test_series_expansion_contract.py
@@ -1,0 +1,118 @@
+"""Tests for the Taylor and Maclaurin series module.
+
+This module contains comprehensive tests for series expansion functionality:
+- Taylor series computation at any point
+- Maclaurin series (Taylor at x=0)
+- Common function series (sin, cos, exp, ln, etc.)
+- Convergence analysis
+- Error bounds
+
+Following TDD and Design by Contract principles.
+"""
+
+from __future__ import annotations
+
+import math
+
+import numpy as np
+import pytest
+
+# =============================================================================
+# SeriesExpansion Class Contract Tests
+# =============================================================================
+
+
+class TestSeriesExpansionContract:
+    """Design by Contract tests for SeriesExpansion class."""
+
+    def test_series_instantiates(self) -> None:
+        """Postcondition: SeriesExpansion can be instantiated."""
+        from signal_toolkit.series import SeriesExpansion
+
+        expansion = SeriesExpansion()
+        assert expansion is not None
+
+    def test_has_max_terms_attribute(self) -> None:
+        """Postcondition: Has max_terms attribute with default value."""
+        from signal_toolkit.series import SeriesExpansion
+
+        expansion = SeriesExpansion()
+        assert hasattr(expansion, "max_terms")
+        assert expansion.max_terms == 50  # Reasonable default
+
+    def test_accepts_custom_max_terms(self) -> None:
+        """Postcondition: Can specify custom max_terms."""
+        from signal_toolkit.series import SeriesExpansion
+
+        expansion = SeriesExpansion(max_terms=100)
+        assert expansion.max_terms == 100
+
+    def test_rejects_non_positive_max_terms(self) -> None:
+        """Precondition: Rejects non-positive max_terms."""
+        from signal_toolkit.series import SeriesExpansion
+
+        with pytest.raises(ValueError):
+            SeriesExpansion(max_terms=0)
+
+        with pytest.raises(ValueError):
+            SeriesExpansion(max_terms=-5)
+
+
+# =============================================================================
+# Taylor Series Method Contract Tests
+# =============================================================================
+
+
+# =============================================================================
+# Maclaurin Series Contract Tests
+# =============================================================================
+
+
+# =============================================================================
+# Get Coefficients Contract Tests
+# =============================================================================
+
+
+# =============================================================================
+# Common Series Functions Contract Tests
+# =============================================================================
+
+
+# =============================================================================
+# Taylor Series Functional Tests
+# =============================================================================
+
+
+# =============================================================================
+# Common Series Functions Tests
+# =============================================================================
+
+
+# =============================================================================
+# Coefficients Tests
+# =============================================================================
+
+
+# =============================================================================
+# Convergence Analysis Tests
+# =============================================================================
+
+
+# =============================================================================
+# Error Bounds Tests
+# =============================================================================
+
+
+# =============================================================================
+# SeriesResult Dataclass Tests
+# =============================================================================
+
+
+# =============================================================================
+# Integration with Signal Toolkit Tests
+# =============================================================================
+
+
+# =============================================================================
+# Utility Functions Tests
+# =============================================================================
