+++ tests/shared/python/signal_toolkit/series/test_series_result.py
@@ -1,0 +1,118 @@
+"""Tests for the Taylor and Maclaurin series module.
+
+This module contains comprehensive tests for series expansion functionality:
+- Taylor series computation at any point
+- Maclaurin series (Taylor at x=0)
+- Common function series (sin, cos, exp, ln, etc.)
+- Convergence analysis
+- Error bounds
+
+Following TDD and Design by Contract principles.
+"""
+
+from __future__ import annotations
+
+import math
+
+import numpy as np
+import pytest
+
+# =============================================================================
+# SeriesExpansion Class Contract Tests
+# =============================================================================
+
+
+# =============================================================================
+# Taylor Series Method Contract Tests
+# =============================================================================
+
+
+# =============================================================================
+# Maclaurin Series Contract Tests
+# =============================================================================
+
+
+# =============================================================================
+# Get Coefficients Contract Tests
+# =============================================================================
+
+
+# =============================================================================
+# Common Series Functions Contract Tests
+# =============================================================================
+
+
+# =============================================================================
+# Taylor Series Functional Tests
+# =============================================================================
+
+
+# =============================================================================
+# Common Series Functions Tests
+# =============================================================================
+
+
+# =============================================================================
+# Coefficients Tests
+# =============================================================================
+
+
+# =============================================================================
+# Convergence Analysis Tests
+# =============================================================================
+
+
+# =============================================================================
+# Error Bounds Tests
+# =============================================================================
+
+
+# =============================================================================
+# SeriesResult Dataclass Tests
+# =============================================================================
+
+
+class TestSeriesResult:
+    """Tests for SeriesResult dataclass."""
+
+    def test_get_series_result_returns_dataclass(self) -> None:
+        """Postcondition: get_series_result returns SeriesResult dataclass."""
+        from signal_toolkit.series import SeriesExpansion, SeriesResult
+
+        expansion = SeriesExpansion()
+        result = expansion.get_series_result(np.exp, center=0, n_terms=10)
+
+        assert isinstance(result, SeriesResult)
+
+    def test_series_result_has_required_fields(self) -> None:
+        """Postcondition: SeriesResult has all required fields."""
+        from signal_toolkit.series import SeriesExpansion
+
+        expansion = SeriesExpansion()
+        result = expansion.get_series_result(np.sin, center=0, n_terms=10)
+
+        assert hasattr(result, "coefficients")
+        assert hasattr(result, "n_terms")
+        assert hasattr(result, "center")
+        assert hasattr(result, "function")
+        assert hasattr(result, "radius_of_convergence")
+
+    def test_series_result_function_is_callable(self) -> None:
+        """Postcondition: SeriesResult.function is callable."""
+        from signal_toolkit.series import SeriesExpansion
+
+        expansion = SeriesExpansion()
+        result = expansion.get_series_result(np.cos, center=0, n_terms=10)
+
+        assert callable(result.function)
+        assert isinstance(result.function(0.5), int | float | np.floating)
+
+
+# =============================================================================
+# Integration with Signal Toolkit Tests
+# =============================================================================
+
+
+# =============================================================================
+# Utility Functions Tests
+# =============================================================================
