+++ tests/shared/python/signal_toolkit/series/test_taylor_series_contract.py
@@ -1,0 +1,152 @@
+"""Tests for the Taylor and Maclaurin series module.
+
+This module contains comprehensive tests for series expansion functionality:
+- Taylor series computation at any point
+- Maclaurin series (Taylor at x=0)
+- Common function series (sin, cos, exp, ln, etc.)
+- Convergence analysis
+- Error bounds
+
+Following TDD and Design by Contract principles.
+"""
+
+from __future__ import annotations
+
+import math
+
+import numpy as np
+import pytest
+
+# =============================================================================
+# SeriesExpansion Class Contract Tests
+# =============================================================================
+
+
+# =============================================================================
+# Taylor Series Method Contract Tests
+# =============================================================================
+
+
+class TestTaylorSeriesContract:
+    """Design by Contract tests for taylor_series method."""
+
+    def test_returns_callable(self) -> None:
+        """Postcondition: Returns a callable function."""
+        from signal_toolkit.series import SeriesExpansion
+
+        expansion = SeriesExpansion()
+
+        def f(x) -> float | np.ndarray:
+            return np.sin(x)
+
+        taylor_func = expansion.taylor_series(f, center=0, n_terms=5)
+
+        assert callable(taylor_func)
+
+    def test_callable_accepts_scalar(self) -> None:
+        """Postcondition: Returned function accepts scalar input."""
+        from signal_toolkit.series import SeriesExpansion
+
+        expansion = SeriesExpansion()
+
+        def f(x) -> float | np.ndarray:
+            return np.exp(x)
+
+        taylor_func = expansion.taylor_series(f, center=0, n_terms=5)
+
+        result = taylor_func(1.0)
+        assert isinstance(result, int | float | np.floating)
+
+    def test_callable_accepts_array(self) -> None:
+        """Postcondition: Returned function accepts array input."""
+        from signal_toolkit.series import SeriesExpansion
+
+        expansion = SeriesExpansion()
+
+        def f(x) -> float | np.ndarray:
+            return np.cos(x)
+
+        taylor_func = expansion.taylor_series(f, center=0, n_terms=5)
+
+        x = np.linspace(-1, 1, 10)
+        result = taylor_func(x)
+        assert isinstance(result, np.ndarray)
+        assert len(result) == 10
+
+    def test_rejects_non_callable_function(self) -> None:
+        """Precondition: Rejects non-callable function argument."""
+        from signal_toolkit.series import SeriesExpansion
+
+        expansion = SeriesExpansion()
+        with pytest.raises(TypeError):
+            expansion.taylor_series("not a function", center=0, n_terms=5)  # type: ignore[arg-type]
+
+    def test_rejects_non_positive_n_terms(self) -> None:
+        """Precondition: Rejects non-positive n_terms."""
+        from signal_toolkit.series import SeriesExpansion
+
+        expansion = SeriesExpansion()
+
+        def f(x) -> float | np.ndarray:
+            return x**2
+
+        with pytest.raises(ValueError):
+            expansion.taylor_series(f, center=0, n_terms=0)
+
+        with pytest.raises(ValueError):
+            expansion.taylor_series(f, center=0, n_terms=-3)
+
+
+# =============================================================================
+# Maclaurin Series Contract Tests
+# =============================================================================
+
+
+# =============================================================================
+# Get Coefficients Contract Tests
+# =============================================================================
+
+
+# =============================================================================
+# Common Series Functions Contract Tests
+# =============================================================================
+
+
+# =============================================================================
+# Taylor Series Functional Tests
+# =============================================================================
+
+
+# =============================================================================
+# Common Series Functions Tests
+# =============================================================================
+
+
+# =============================================================================
+# Coefficients Tests
+# =============================================================================
+
+
+# =============================================================================
+# Convergence Analysis Tests
+# =============================================================================
+
+
+# =============================================================================
+# Error Bounds Tests
+# =============================================================================
+
+
+# =============================================================================
+# SeriesResult Dataclass Tests
+# =============================================================================
+
+
+# =============================================================================
+# Integration with Signal Toolkit Tests
+# =============================================================================
+
+
+# =============================================================================
+# Utility Functions Tests
+# =============================================================================
